Struct git::fs::GitFS

pub struct GitFS { /* fields omitted */ }

default structure used to contain some information regarding the git repository some information such as the file path.

Methods

impl GitFS

fn new(p: &Path) -> Result<Self>

Open a git from the given path (the git directory must be valid)

Will fail if the given directory does not look like a valid git directory.

Example

use std::path::*;
use git::fs::GitFS;
let git_path = PathBuf::new().join("path").join("to").join("nowhere");
match GitFS::new(&git_path) {
  Ok(_)    => { /* I have a valid git repository */ },
  Err(_) => { /* invalid path or invaild git repository */ }
}Run

This function will open the directory and check it is a valid repository and then returned the loaded GitFS.

TODO: rename to open

fn refs_dir(&self) -> PathBuf

return the refs directory (where all the link to the branches and tags are)

Example

use std::path::*;
use git::fs::GitFS;
let git_path = PathBuf::new().join(".").join(".git");
let git = GitFS::new(&git_path).unwrap();
let refs_dir = git.refs_dir();
assert_eq!(refs_dir, PathBuf::new().join(&git_path).join("refs"))Run

fn objs_dir(&self) -> PathBuf

return the objects path (contain all the blob and packed git objects)

fn info_dir(&self) -> PathBuf

return the info directory path

fn hooks_dir(&self) -> PathBuf

return the hooks directory path

fn config_file(&self) -> PathBuf

return the git config file path

fn description_file(&self) -> PathBuf

return the git description file path

fn head_file(&self) -> PathBuf

return the git current HEAD file path