Enum git::object::TreeEnt

pub enum TreeEnt<H: Hash> {
    Tree(Permissions, PathBuf, TreeRef<H>),
    Blob(Permissions, PathBuf, BlobRef<H>),
}

the different type of entity managed by our current implementation

Tree: reference with a permission to a sub tree (recursive entry). This is equivalent to a filepath directory.

Blob: reference with a permission to a blob of data. This is equivalent to a file.

Variants
Tree(Permissions, PathBuf, TreeRef<H>)Blob(Permissions, PathBuf, BlobRef<H>)