Struct git::object::Blob

pub struct Blob(_);

Blob data

So far this is only in-memory data but it should become something more efficient in near future.

A Blob is referenced by a BlobRef<H>.

Discussion

This part is still under construction and the API might (certainly) change. So far we are storing everything in memory (a Vec). It is does not make much sense to do so as we could blow the memory and be quite slow to process the data.

Ideally, in the future, the Blob may become a trait so we could use streamable objects or in memory data depending on what is better.

The composition of a Blob may also differes depending of the backend in use. So far we will use in the filesystem as it is the legacy one but ideally we could change the backend to a key-value database which won't be any different.

Methods

impl Blob

fn new(data: Vec<u8>) -> Self

create a blob from the given data.

fn as_slice(&self) -> &[u8]

access the inner data as an immutable slice of bytes