Struct git::object::Date

pub struct Date(_);

Git date object

Based on the chrono crate.

It is always at the current Local time, without nanoseconds.

Example

Creating a new date for now:

use git::object::Date;

let now = Date::now();
println!("Today is: {}", now);Run

Creating a custom date:

use git::object::Date;

let date = Date::ymd_hms(2013, 1, 29, 15, 34, 11)
            .expect("to have a valid date and time");
println!("First steps in UK: {}", date);Run

Methods

impl Date

fn now() -> Self

create a new date set to know and the system local timezone

use git::object::Date;

let now = Date::now();
println!("Today is: {}", now);Run

fn new(dt: DateTime<Local>) -> Self

create a new date from the given DateTime

This function will filter out the nano second precisions (if any).

fn seconds_since_epoch(seconds: i64) -> Self

create custom time from seconds since epoch (using local timezone)

use git::object::Date;

let date = Date::seconds_since_epoch(1479973175);
println!("that day: {}", date);Run

fn ymd_hms(year: i32, month: u32, day: u32, hour: u32, minutes: u32, seconds: u32) -> Option<Self>

Convenient function to make up date from human logic

use git::object::Date;

let date = Date::ymd_hms(2013, 1, 29, 15, 34, 11)
           .expect("to have a valid date and time");
println!("that day: {}", date);Run

fn encode_for_obj(&self) -> String

encode in a string for the git object format

i.e.: Seconds since EPOCH followed by timezone (+/-HHMM).

use git::object::Date;

let date = Date::ymd_hms(2016, 11, 24, 7, 39, 35)
           .expect("to have a valid date and time");
println!("that day: {}", date.encode_for_obj());Run