Struct git::object::Person

pub struct Person { /* fields omitted */ }

Git's Person data type

Base type for committer and author (embedding a date along):

a full name;

an email address;

a date (the date associated to when the person did something on the current object).

Enhancement

Date and Person

Semantically, it does not make sense to define a Person with its actions logged along. A Person's is not defined by when it actually did something. However, in the context of git, a Committer and an Author will have the same data and will be tagged along a date.

Data Validation

There is not validation of the username or the email address. This is customer's responsabilities to store valide data (or not).

Example

use git::object::Person;

let me_now = Person::now( "Nicolas".to_string()
                        , "my@email.address".to_string()
                        );
println!("Hello! Right now, I am: {}", me_now);Run

Methods

impl Person

fn now(name: String, email: String) -> Self

convenient function to create a Person with the current Date

use git::object::Person;

let me_now = Person::now("nicolas".to_string(), "my@email.address".to_string());
println!("{}", me_now);Run

fn new(name: String, email: String, date: Date) -> Self

create a new Person

use git::object::Person;
use git::object::Date;

let now = Date::now();
let me_now = Person::new("nicolas".to_string(), "my@email.address".to_string(), now);
println!("{}", me_now);Run

fn name(&self) -> &str

access the Person's name

use git::object::Person;

let name = r"nicolas";
let email = "my@email.address".to_string();
let me_now = Person::now(name.to_string(), email);
assert_eq!(name, me_now.name());Run

fn email(&self) -> &str

access the Person's email address

use git::object::Person;

let name = "nicolas".to_string();
let email = r"my@email.address";
let me_now = Person::now(name, email.to_string());
assert_eq!(email, me_now.email());Run

fn date(&self) -> &Date

access the Person's date

use git::object::Person;
use git::object::Date;

let name = "nicolas".to_string();
let email = "my@email.address".to_string();
let date = Date::now();
let me_now = Person::new(name, email, date.clone());
assert_eq!(&date, me_now.date());Run