Trait git::protocol::Hash

pub trait Hash: Sized {
    fn hash<R: BufRead>(data: &mut R) -> Result<Self>;
    fn from_bytes(Vec<u8>) -> Option<Self>;
    fn digest_size() -> usize;
    fn as_bytes(&self) -> &[u8];

    fn from_hex(s: &str) -> Option<Self> { ... }
    fn digest_hex_size() -> usize { ... }
    fn to_hexadecimal(&self) -> String { ... }
    fn decode_bytes(i: &[u8]) -> IResult<&[u8], Self> { ... }
    fn encode_bytes<W: Write>(&self, w: &mut W) -> Result<usize> { ... }
    fn decode_hex(i: &[u8]) -> IResult<&[u8], Self> { ... }
    fn encode_hex<W: Write>(&self, w: &mut W) -> Result<usize> { ... }
}

Hash Protocol

Originally, git has been using SHA1 to generate unique identifier (ref) for objects, commits. In order to prepare the field for a other Hash algorithm (but also to be able to mock when testing) we provide a protocol for hashing.

Common interface

The idea is to provide a simple interface so we can easily change/customise the hashing mechanism later on.

Example

use git::protocol::{SHA1, Hash};

let data = "hello world";
let hash = SHA1::hash(&mut data.as_bytes()).unwrap();
assert_eq!(hash.to_hexadecimal(), "2aae6c35c94fcfb415dbe95f408b9ce91ee846ed");Run

Required Methods

fn hash<R: BufRead>(data: &mut R) -> Result<Self>

function to hash a stream

fn from_bytes(Vec<u8>) -> Option<Self>

fn digest_size() -> usize

the size of the digest

fn as_bytes(&self) -> &[u8]

Provided Methods

fn from_hex(s: &str) -> Option<Self>

fn digest_hex_size() -> usize

the size of the digest in hexadecimal

fn to_hexadecimal(&self) -> String

return exadecimal encoded of the digest

fn decode_bytes(i: &[u8]) -> IResult<&[u8], Self>

fn encode_bytes<W: Write>(&self, w: &mut W) -> Result<usize>

fn decode_hex(i: &[u8]) -> IResult<&[u8], Self>

fn encode_hex<W: Write>(&self, w: &mut W) -> Result<usize>