Trait git::protocol::Repo

pub trait Repo {
    fn is_valid(&self) -> Result<()>;
    fn get_description(&self) -> Result<String>;
    fn get_ref<H>(&self, r: SpecRef) -> Result<Ref<H>> where H: Hash;
    fn lookup_hash<H: Hash>(&self, prefix: &Partial<H>) -> Result<Vec<H>>;
    fn get_object<H, O>(&self, r: O::Id) -> Result<O> where H: Hash, O: Object<H>, O::Id: Hash;
    fn get_object_<H>(&self, r: H) -> Result<Obj<H>> where H: Hash;
    fn list_branches(&self) -> Result<Vec<SpecRef>>;
    fn list_remotes(&self) -> Result<Vec<SpecRef>>;
    fn list_tags(&self) -> Result<Vec<SpecRef>>;

    fn get_ref_follow_links<H>(&self, r: SpecRef) -> Result<H> where H: Hash { ... }
    fn get_commit<H: Hash>(&self, h: CommitRef<H>) -> Result<Commit<H>> { ... }
    fn get_tree<H: Hash>(&self, h: TreeRef<H>) -> Result<Tree<H>> { ... }
    fn get_blob<H: Hash>(&self, h: BlobRef<H>) -> Result<Blob> { ... }
    fn get_object_ref<H>(&self, r: Ref<CommitRef<H>>) -> Result<Commit<H>> where H: Hash { ... }
    fn get_head<H: Hash>(&self) -> Result<Ref<H>> { ... }
}

Required Methods

fn is_valid(&self) -> Result<()>

common function to validate the given Git Repository is valid. See GitFS.

fn get_description(&self) -> Result<String>

get the description of the given Git Repository

fn get_ref<H>(&self, r: SpecRef) -> Result<Ref<H>> where H: Hash

standard function to get the Hash associated to a given reference (HEAD, master... remote/master...)

fn lookup_hash<H: Hash>(&self, prefix: &Partial<H>) -> Result<Vec<H>>

fn get_object<H, O>(&self, r: O::Id) -> Result<O> where H: Hash, O: Object<H>, O::Id: Hash

get object from a given hash ref

fn get_object_<H>(&self, r: H) -> Result<Obj<H>> where H: Hash

fn list_branches(&self) -> Result<Vec<SpecRef>>

fn list_remotes(&self) -> Result<Vec<SpecRef>>

fn list_tags(&self) -> Result<Vec<SpecRef>>

Provided Methods

fn get_ref_follow_links<H>(&self, r: SpecRef) -> Result<H> where H: Hash

follow the links for a given Ref until a HashRef. HEAD -> master -> abcdef012345678..

This function is a combination of get_ref and patten match on the Ref enumeration.

fn get_commit<H: Hash>(&self, h: CommitRef<H>) -> Result<Commit<H>>

fn get_tree<H: Hash>(&self, h: TreeRef<H>) -> Result<Tree<H>>

fn get_blob<H: Hash>(&self, h: BlobRef<H>) -> Result<Blob>

fn get_object_ref<H>(&self, r: Ref<CommitRef<H>>) -> Result<Commit<H>> where H: Hash

default implementation to read an object (a commit if Ref is a SpecRef) from a given Ref.

This default implementation is a combination of get_ref_follow_links, pattern match and get_object.

fn get_head<H: Hash>(&self) -> Result<Ref<H>>